Spinner for inputs
For loading inputs, the `aria-busy`-attribute needs to be applied to the wrapping `label`, because replaced DOM elems, like inputs, doesn‘t support pseudo elems in most browsers. (See [the W3C reccomendation](http://www.w3.org/TR/2005/WD-CSS21-20050613/generate.html#before-after-content) for more on this.)

The spinner is applied as an absolutely positioned pseudo elem, so the label text needs to be `nowrap`ed.

References #138

Signed-off-by: Jørgen Arnor G. Lom <j@lom.me>