fix: Changed RelativeURL --> relativeURL in BizHttp.ts
Because we get capitalized camel case variables in the classes we
get from the backend, but in the classes that are created in the
frontend we use lower case camel case on the variable names and
capitalized camel case on the classes.

The main reason for doing this change now is because BizHttp didn't
work the way it was.