feature(Core): Removed withoutJsonMap parameter from unihttp.send()
This makes it easier to read and understand what is going on and gives
more control to the developer.

Because it is hard to know what the argument means when reading the
code. And if you just see the method signature it is hard to know that
it just gives you JSON if you use .send({}, true), and the response
object if you use .send().

Plus double negatives are super hard. For example what does this result
in:

    var foo = false;
    var withoutJsonMap = !foo;
    req.send({}, withoutJsonMap);

If you had to look at the examples in the second paragraph you have
already lost :-)