feature(Accounting): Made the routes to trasquery detail view generic
Because we want to link to this view form anywhere and don't want to
have to make specialized routes for every query we want to make. So now
we can just use angular2's query parameters to set the odata fields.

The reason we don't start using the `routerLink` directive is because we
are having trouble making it work in the unitable template function.
Probably because it injects it too late in the angular 2 life cycle.

The url notation we use for the query parameters are something called
"Matrix URL notation", you can read more about it in the docs:
https://angular.io/docs/ts/latest/guide/router.html#!#query-parameters

ref: #909