feature(Accounting): Added logic to the VAT report view
Converted the two sub views into components because the dependencies of
flow and data was too big to have it as separate views. I could delete
lots of code when I converted them to views. And we download all data up
front anyway because the summary per post we show as default and the
control view we download the data to show the mini checklist in the to
right corner.

Added UniCurrencyPipe directive to have one place where we convert the
currency numbers in the app to one standard format.

Converted the query parameters to the transquery details view to replace
underscores(_) with dots(.) after it receives the field names because
lite-server doesn't work with dots anywhere in the url. That is because
it uses the dot in the url to decide if it is a frontend url or if it is
a file it is supposed to serve.

closes: #890
closes: #891
closes: #892
closes: #909