fix: Fixed a bug where moment/locale/nb.js could not find moment
Had an issue where a locale file (moment/locale/nb.js) for moment.js
could not find moment js (tried to get ../moment.js) because we use a
minified version (moment/min/moment.js). Added a stream map to convert
the moment file path to moment/moment.js.

This worked before in some browsers and we thing that was because we
still had the global moment instance. It would most likely break
everywhere when we removed that.

This change has been peer programmed with @urrang