fix(Sales): View unresponsive after implementation of new router
Because the new router re-uses views when only the url-parameter
changes. That means that what was in the onInit before must now be put
in the route.params.subscribe() function.

Since I edited most of the getLayoutAndData() functions anyway I took
the time to rename all of them to setup() since they get don't get
(return) anything.

closes: https://github.com/unimicro/AppFramework/issues/1658