feature(Accounting): Add button to fetch receipt from Altinn
The reason for re-writing the logic for getting pin code message from altinn was that the old logic was so hard to understand because it re-used the same object and endpoint for all the communication for the two factor authentication. What remains is an extra check for see that the userID/password/pin is valid before closing the modal. But we have to do this at a later stage because altinn doesn't validate userID/pass when we ask for which pin to use in the test environment.

closes: #1059