refactor(uniform): count elements inside ready handler
New release candidate RC5 has some changes respect how ViewChildren works
It is affected by ngIf directive.
In case some element is showed by ngIf asynchronously it will not be available on
the viewchildren component until it is painted.
For this reason some times elelements of uniform are not available at ngAfterViewInit
so we can't calculate the proper number of elements.
For that reason we have moved that count to the readyHandler

ref: #1141