fix(Accounting): Prevent overriding VAT report Altinn comment
Because we had an issue where the user wrote something, then waited
400ms, then wrote something more. When he waited the http request was
sent, and what ever he wrote after the http request was sent was
overridden when the request came back.

Now we ignore the returned vat report from the backend and always keep
what the user wrote in the textarea.

ref: #1169