Add unientities.ts to the git repo
Because we want to have reproducible builds/commits that will be the
same the next time we run them. This will also give us traceability to
see who added what to the unientities.ts file. If we break something
in the unientities.ts file it will only affect the person waiting for
something new in the file, not everybody else. And finally we will be
able to get meaningful diffs from the unientities file.

This is something that have to be done sooner or later no matter what
because we will get more parallel branches (production, staging,
development and so forth) that need different versions of the
unientities file.