fix(Core): Change form controllers to match FieldLayoutDto(unientitites)
In the controllers the placeholder has been expected to live under the
Options object, but according to FieldLayoutDto it should live directly
on the files object. I have now fixed that.

I also took the opportunity to fix all the places where placeholder was
accessed on "placeholder" instead of "Placeholder" which is what
FieldLayoutDto says it should be called.