feature(Accounting): Added new button and account group to account
Because the user sometimes wants to make new accounts...

I added <UniFieldLayout> in front of all the form filed layouts so that
it won't brake in the future if they change the types in the backend. I
probably also should have removed the default settings on the files,
because they don't actually _need_ to be specified. But all of this goes
out the window when we get it from the backend so I'll just let it be.

closes: #1292