feat (uniquery): add components and services for new uniquery module
Quite a lot of changes, because this will be a really cool module.. added components, styling, and services used to generate uniqueries (dynamic reports). The user is also able to save/edit queries and to export data to Excel. UniHttp changed to make it possible to get Blob content type (experimental feature in Angular 2)

New routes and ngmodule added after merge with @jorgeunimicro's changes yesterday

The module uses UniTable to retrieve/display the data. UniTable updated to new version to support the features needed in uniquery

closes #1327