fix (altinn): quickfix to make altinnauthenticationdatamodal work again
Because of changes in RC6, the component within the modal does not seem to be created until it is actually opened. In this modal, we tried to get data from it before it was opened. I have added a quick fix that opens (and sometimes closes) the modal. I think the code needs to be refactored more to get a permanent workaround, but I'll leave that to @karl-gustav.

@sondrevik this hopefully solves your problem for now

ref #1380