fix(Accounting): Moved the AccountNumber change listener into http cb
Because when the response from getting the account numbers runs
this.extendFields() which in turn runs _.cloneDeep on this.fields,
which, in turn, destroys any bindings we have into fields in the
original this.fields array.

Therefore we moved the setting of the subscriber into the callback for
the accounts call, and put it into a setTimeout so that it runs after
unform is done setting up the fields again.

closes: #1381