feature(Core): Add yarn.lock to be able to use yarn instead of npm
We want to try out yarn because we want to have deterministic builds.
I.e. builds that are the same on all machines. This is not something
currently get from npm, you can try and use shrinkwrap to acheive
something similar, but it is not as good as yarn. Yarn on the other hand
has this functionality built in from the start.

The folks at NPM seems to welcome the new package manager:
http://blog.npmjs.org/post/151660845210/hello-yarn

Here is a link to an overview of the yarn commands and how they relate
to the npm commands:
(Summary: use `yarn add <pkg>` instead of `npm install --save <pkg>`)
https://yarnpkg.com/en/docs/migrating-from-npm