Feature/jff/angular rc to final release #1186 (#1486)
refactor(app): update angular to version 2.10

Update angular to version 2.1.0

Those changes make our application compatible with the last changes of angular. That means that the application use now NgModule, and all our components and routes are grouped in that way.

Also a problem with RxJS was found during the updating. Rx bundle that is in node_modules doesn't work with this version of angular. It is referenced in angular's issues. To fix that we have bundled RxJS libraries using a gulp task. That should be a provisional solution and that should be fixed in the future. When that moment comes we should remove that task from gulp.