feature(Core): Add submit on enter to the altinn pin form
Because we want to be able to press enter for submitting to speed up the
usage of the form.

As a side note I would like to say that this sort of goes against the
usage of enter as a navigation key. But life is short and there might
come changes to uniform that handles this for us so I won't use any time
on it.

closes: #1440