Complex statustrack (#1434) (#1561)
Added support for more complex status tracks, implemented on A-melding.

The IStatus-interface now has an substatusList-array of other IStatuses, which allows you to put sub-statuses on your statuses. It's also expanded to contain a timestamp-field (which is shown on substati) and a data-field which can contain any arbitrary data you might want to keep, in order to identify the status later on.

Clicks on a status now also emits an statusSelectEvent, which passes the clicked status, as well as the parent, if it's a substatus. (This is output on both the uni-statustrack and the uni-toolbar-component, so you can use it seperately, or as part of the toolbar.)