feature(Core): Added different configs for dev, prod, and pilot
Because we want to be able to build for different environments on the
jenkins server. This also means that we need to take into account what
environment we run the `gulp entities` task for, so I deleted the
default url and the users are now forced to manually specify the url
before they run the gulp entities task.

The reason this works out of the box is that for some time we have
committed the unientities.ts file into git.