feature(UniQuery): Add UniQuery to the customer view
Added a new model interface for id/name objects because when we use the
statistics endpoint we will probably often get that kind of data.

Changed the URL to the customer endpoint from `/customers/list` and
`/customers/details/:id` --> `/customers` and `/customers/:id` because
that is the way we want to have our URL schema. This is just a
continuation of the work @urrang started.

The reason for not using the URL router for navigating to the UniQuery
excerpts is that we want to keep the queries in the DOM for caching
purposes. The user will navigate a lot between these views and we don't
want to re-render them each time. This is also the reason for the
complex change detection in the UniQueryReadOnly component.

ref: #1390
ref: #1558