fix (accounting): improve keyboard handling pro journalentry grid
Enable copy/paste of account/vattype to other rows and add "+" as a shourtcut for adding a new journalentry

because this is what people are used to in V3 and it improves flow when registerring multiple journalentries

Also refactor calcfunction for journalentries locally instead of running all calculations in the API. Also run validations in frontend - the data will also be validated in the API when trying to save it. This is done to make the view more "snappy" and because it isn't really that hard to do the math in appfrontend (... though I might regret saying this..)

Also add extra column in pro grid for net amount. This is useful to show the user what calculations are done on a line-level. It also enables the user to input the net amount instead of the gross amount. Needs testing from BA to make sure this works as expected, but looks good to me. 

closes #1641