feat (sales): add lists and more data to customer and supplier details
because we want do display e.g. quotes on a customer and bills on a supplier. This also relates to a commit in the API to add migrations for these lists

also uniquery is extended to support fieldtype to better display numbers and dates, and statuscode is automatically converted to the status description from the API (using mapping in the frontend).

also do some cleaning up on customer and supplier to display error messages better and use the toolbar component in supplier

closes #1684
closes #1651
closes #1650