refactor(salary): remove usage of staticregisterservice
this services causes to much load on the localStorage. Use statistics instead. localStorage was only used for styrk/postalcodes in a couple of components as far as I could tell

I have not deleted the service from appfrontend, but disabled retrieving data for now. It could probably be removed, but if so it should also be removed from the API, so I'll leave it for now.

closes #1738