feature(Common): Set autofocus = true on "all" UniForm instances
Because we want to minimize the users need to use the mouse.

The reason for not putting it into PayrollRunDetails and WageTypeDetails
is that there where some issues getting the autofocus to work and
therefore we abandoned those views for another day when we are not the
day before a major demo.

closes: #1763