fix(Accounting): Trim Leverandørreskontro & Kundereskontro from account
Because we don't have the space to show both account number and the full
account name. This is just a quick fix, I have created an issue in the
backend to fix this properly.

Code review done by showing the code/solution on my computer to
@kjetilek.

backend issue: https://github.com/unimicro/AppFramework/issues/2177

closes: #1821