fix(Core): Add common error handling for the entire app
Because we badly need a common way of handling errors. Both because of the fact that we were missing error handling in a LARGE percentage of the app, and the fact that it was implemented vastly differently tough out the app.

The common error handling does three things:
1. Send error to Raygun
2. Send error to console
3. Show a toaster to the user

Look at this commit for examples of how to use it.