Feature/sev/payroll run caching #1377 (#1884)
* refactor(PayrollRun): move save up to payroll run

moved save up to payroll run, making it responsible for all saving
moved filter up to payroll run, so we can use filter on employees

* feature(PayrollRun): manage saving

managed saving and calculating
cached wagetypes

* feature(SalaryTransaction): remove unused code

all saving and getting is now handled by payrollrun

* refactor(SalaryTransaction): moved detail header to selection list

with summary. Making salarytransList more focused on salary transactions

* fix(PayrollRun): making section work again

after going over to caching, the section didn't work as expected
removed canPost, because it made posting modal dissappear when we
gave payroll run booked status.
made salary trans table not editable when saving

* fix(SalaryTransactionSupplement): make fields readonly on status > 0

because we don't wan't to change on a calculated payroll run

* refactor(SalaryTransaction): replace dirty transes with new ones

now we can replace transes where they are in the list. This way they don't
shuffle around when saving

* fix(PayrollRun): fixed toggle on section

now we actually get alerted when description is missing
and selection table didn't focus on first when starting in form
so i set focus every time we toggle from payroll to salary transactions

* fix(PayrollRun): setting delete button when payroll run is here

this way we can't delete if status is calculated or higher
also, we hide employee arrows when we don't have any employees

* fix(VacationPayModal): made vacation pay modal update salary transes

so we can see that the new transes has been generated

* fix(PayrollRun): removed caching from payroll run service

because UniView handles this much better

* fix(SalaryTransaction): make delete ignore recurring post

because the way we did it before did not work on recurring transes
now we setup delete with rules if we can delete, or disable delete
if status is calculated or higher

ref #1377