Feature/sev/salary trans fixes #1889 (#1956)
* fix(SalaryTransaction): made it more robust

maybee this is the reason for undefined of undefined error

ref #1889

* fix(PayrollRun): nullcheck on selection list

ref #1889

* fix(SalaryTransaction): giving value to expanded fields after save

ref #1889

* fix(SalaryTransaction): updating sums

now payroll run update sums in selection list when we get
new salary transactions

ref #1889

* fix(SalaryTransaction): making it possible to zero out autocomplete

ref #1889

* fix(SalaryTransaction): added expands

so that we always have something in the expanded fields. Now we can
press enter in autocomplete fields and get the value on the row above
also, we only call update row if we get rate. if we don't, it get's
updated when we return row in changeCallback

ref #1889