fix (accounting): use on-demand lookups for accounts in journalentry
because the caching is too slow when lots of customers/suppliers have been created. Also lookups are now way faster than before, so it doesnt feel slow when lookup is run on demand

also change to use statistics to get accounts - map to regular objects before returning to component. This will reduce the amount of data retrieved a lot

closes #1969