Added crossOrigin anonymous to all scripts loaded with systemjs
Because we want to try and prevent the generic "script error" error
message in Raygun.

It is hard to test if this works locally or in development since we
don't use Raygun there, therefore we have only tested that this doesn't
break existing functionality and we'll test it more throughly in the
pilot environment.