fix(Dashboard): Dashboard doesn't update on client change
Because we didn't reload the view on this.authService.companyChange
event. And we need to do that because this view is not re-initialized
when we change client, it is re-used as it is the only view open on a
new client. Unless of course it is closed when we change client, then it
is re-initialized.

closes: #1977
code review: #2028