fix(accounting): improve attachmenthandling journalentry (#2031)
Allow user to upload file before selecting journalentrynumber. Also remember if the user uses attachements when opening the journalentryview (using localstorage).

Because a user will normally either always use files, or don't use files, so it makes sense to remember it. Also the process will often start with uploading a file, so it does not make sense that the user has to select the journalentrynumber first