fix(Accounting): Fix journal entry lines missing from transquery view
Because we had issues with the join that was added for showing if a line
had attachments. @kjetilek made some changes in the backend to accommodate
the new query, but it is only a quick fix, it'll have to be fixed more
properly later. Have a look at the github issue for more info.

The accountDetailsReport view uses the same logic so added the same fix there.

Also added some better error propagation from the statistics endpoint
service because when we don't fail fast the app crash will happen far
from where the error occurred.

closes: #2016
code review: #2060