Moving uni-save to the toolbar (#2075)
We've moved uni-save to be part of our toolbar. The interface for statuses and save actions are still the same, but instead of passing it to a <uni-save>-element directly, you now add it to the toolbar config object.

Mostly this'll just be search-and-replacing this.saveactions => this.toolbarconfig.saveactions, and deleting the old <uni-save> component from your template.

I've implemented the new config on the Payroll runs and Employee details GUIs, and a super sexy deprecation warning on the old uni-save implementation.