feature(Unidate): Set LocalDate as a date type in AppFrontend
We need to be able to send dates as "2016-04-04" to the backend
because of our problems with Javascript's Date type (which in reality is
a DateTime). The problems stem from the fact that time is a tricky
thing if you are working on an international server, it changes from
where you are in the world.

`2016-01-01T00:00:00+0100` becomes `2015-12-31T23:00:00Z` on
the backend, and if you remove the time part from this string to check
the date it is wrong. Therefore we send `2016-01-01` which is implicit
`2016-01-01T00:00:00Z` and therefore can be done before we have
actually implemented any Date (as oppose to DateTime) on the
backend.

But if we only use date it is the same everywhere. Think of it like
Christmas eve, (24th of December), the date might start on different
times (relative to UTC) around the globe, but everybody agrees that it
is on the 24th of December.

closes: #2114, #2115
code review: #2113