fix(LocalDate): Fix LocalDate implementation and update uni-form/table
Update uniform and unitable because their versions of LocalDate was
lagging behind.

Update LocalDate because we want to be able to easily create a new
moment date based on a LocalDate. Now that can be done as easily as
`moment(new LocalDate('2016-04-04'))`. This functionality already worked
for date: `new Date(new LocalDate('2016-04-04'))`