hotfix(OneSignal): Disable OneSignal in pilot
We are now calling
https://devapi-unipushadapter.azurewebsites.net/api/devices directly
from pilot.unieconomy.no and that call is failing because the CORS
settings for that server doesn't have https://pilot.unieconomy.no listed
as a valid server.

But that is kinda besides the point, we should in no way be calling
dev* anything from the pilot environment.

Also changed the prod config to prevent this kind of problems when we
get a production environment.