feat(accounting): handle financialyears better and add calc in journalentry
because we should validate and warn the user if he/she registeres data on several financialyears, and we should use the current financialyear when getting the journalentrynumbers

also add calculation on current row (opening balance, changes, vat, etc) because this is usefull information for the user. Build the solution based on the way it looks in V3 and input from Rune. Additional styling will be added by @jlom later, current styling is based on setup in V3

closes #2260
closes #2266