feat(accounting): improve reconliciation and transquerydetails
transquerydetails: because the transquerydetails was not effective to work with compared to the "Forespørsel på konto" in V3. Add new filtering in the top of the page based on input from Svein. Also improve styling (colors), and remember columnvisibility. Also show tooltip for Description column, because this could be quite long.

reconciliation: make adjustments based in input from Sveinung in unimicro/AppFramework#2563

closes unimicro/AppFramework#2563
closes #2293