fix(accounting): dont include journalentries for next year in summary
the journalentrymanual view retrieved information about accounts as the user added new accounts. These numbers included the total sum for the account if it was a balance account, and the sum for the current year if it was an result account. This was misleading if you added journalentries on a balance account the next year, so we now filter out period.accountyear le current-year for balance accounts

closes #2370