Just call focus if we have more than one company
When we login in first time in the app and we just have one company the component tries to navigate directly to the dashboard without render the company select. Despite the fact the select is not rendered the app tries to set the focus. For that reason now we just set the focus when the select is renderend, in other words, when we have more than one company to select.