refactor(accounting): translate journalentrydata objects to/from draftlines in frontend
because this gives us more flexibility now that we are going to implement even more features in the journalentryviews (currency, accruals, payments, etc). Use the API code to "translate" to the same Typescript functionality, and consider removing the backend functions when that is complete.

Also refactor frontend to use LocalDate because this was a good time to do it (because of the refactoring of journalentrymapping), and it had to be done at some point. Most entities related to economy/sales are updated, this also relates to (much larger) API change in unimicro/AppFramework#2772

closes #2575