feat(ticket): start implementation of ticker module
because we need a simple way to create new views/lists that can be used to find data. Default lists are configured through a json-file for now, this might be moved to an API later.

frames for solution is implemented in this commit - design is based on input from design team. Not all functionality is completed yet, but several components are set up and ready for further implementation. More information about what will be available in the end is available in #2635 

For now the views are not available through the menu, just by knowing the routes, "/tickers"

closes #2643
closes #2652 
ref #2645 
closes #2653 
ref #2658