fix(salaryRoutes, index): re-add UniSalary to routes and fix index.html
UniSalary was removed during webpack refactor since it doesnt appear to
do anything meaningful. Turns out it was used as a selector for quite
a bit of styling in the salary components. Re-adding it as it might also
be used as a "salary dashboard" in future iterations.

Also re-added appLoading css class to the text that is displayed while
loading frontend. This was removed by accident while fixing webpack bugs