feat(ticker): add support for printactions to ticker and implement ticker-wrapper to simplify ticker usage
because this makes it easier to use tickers standalone, i.e. to support the more standard invoice lists

also add support for action overrides, to support more complex actions when using the ticker-wrapper (e.g. payments, that needs input from user). Documented in confluence how this works, and also implemented invoicelist because this is one of the most complex lists

closes #2883
closes #2884
ref #2654