fix(Unisearc): Fix the bug when you write a number in a customer search
Because we got this error message:
ORDER BY items must appear in the select list if SELECT DISTINCT is
specified.

And the reason we got that is because we change the field we sort by if
we search for a number, because we assume that the user is searching for
a CustomerNumber if they write a number. And when that field is not in the list of selected fields we get the error message.