feat(accounting): add invoicenumberhandling in journalentry
because this allows the user to use journalentry to register journalentries for invoices (both supplierinvoices and customerinvoices) without actually using the bill/customer invoice views. This is useful e.g. when using other systems for invoicing.

also this makes it easy to register payments in/out using the journalentry. A modal is displayed if more than one journalentryline with the specified invoicenumber exists - if not we just assume that the existing item is what the user wants

the functionality is pretty much the same as in V3

closes #2984