fix(settings): improve bankaccounthandling in settings
because it didnt work quite as expected always. Maybe it still doesnt - difficult to know what to expect from this component - but at looks better now. Change the account lookup to only get data on demand instead of preloading it like we used to do - this also makes the component faster to load, and improves the user experience

closes #2444