fix(accounting): implement agiohandling for postpost markings
because this is expected behaviour and avoids problems if we get good data into the system when registering it, instead of trying to clean it up later. Do this when setting invoicenumber manually as well (i.e. not using Innbetalinger), and fix some problems in the Innbetalinger view (if you wanted to partially pay an invoice that was not possible - also fix some rounding problems)

add confirm modal if diff in percent between currency rates used in the agio dialog are larger than 15%, to avoid problems with partial payments when only the currencyamount is changed

also add block for using the Åpne poster view (ledgeraccountreconciliation) to mark rows with different currencies, this causes problems in the API so it shouldn't be accepted in the appfrontend either.

also remove deprecated component transquerylist because this has been replaced by a better solution and is no longer needed (it hasnt been used for a long time, but the component had not been deleted)

ref #2769