feat(accounting): add typeahead component to journalentry description
because this makes it easy for the user to enter predefined texts, e.g. use code 1 to insert Innbetaling. The hardcoding of the options is temporary - @trondjp is working on getting these texts from the server, so when that is done this component should be changed to get the texts from the API

ref #3144