feat(feedback): move feedback component to layout module and navbar
because it takes up too much space in the bottom of the screen now, and overlays data the user might want to see. Also switch to using toastservice for success feedback, because that makes more sense now (might not have been implemented when the feedback component was implemented). 

Might need some styling from @jlom, but it seems to work as it should now

closes #3252