refactor: move uniform, unitable and unisearch to AppFrontend. Update angular to version 2.4.2
Move external repos (UniForm, UniTable and UniSearch) to appfrontend. That will make easy to debug those complex components against AppFrontend.  Before it was too hard to debug errors in AppFrontend that where related with those repos but were impossible to reproduce them out of the box. With this change we expect that improve and increase core components will be easier.