fix(Employee): get employee to save correctly on navigation (#3454)
because we cleared cache before we got the chance to save from the cache.
removed all ekstra subscribes and used .switchmap and .do instead. This Way we can figure out when saving is done and then clear chache. Also this can happen in the background, so the user leaves the view and saving actually happens in the background(Something we thought happenen, but didn't because of the clear cache problem)

UL-182