feature(Editorconfig): Add .editorconfig file to project
Because we have issues with people using different editors and different
rules for auto formating their code, this results in lots of hard to fix
merge conflicts.

This is not a complete solution, this only fixes charset, indentation,
and trailing whitespace. VS code for instance likes to insert spaces
around imports in ts files, this is potentially a lot of lines of
whitespace changes in a big file.

.editorconfig should work out of the box with all of the editors I know
of that people in this project uses (except Atom). For a full list see:
http://editorconfig.org/#download