feature(Bureau): Add bureau list
Because we want a way for the super users to have a quick overview of
what companies they have access to.

This was originally started by @urran and finished by me, so the widget
changes are from @urrang. This is still missing the actual values of
waiting tasks in the different companies, but that will come in the next
iteration. For now we just really need the button for creating new
companies since that functionality is going away from where it is
currently implemented.

ref: UC-153