fix(uniform text): revert changes made to blurHandler a few days ago
Because they led to a behavor where if you didnt specify a template
on the field, the change event would always emit undefined,
regardless of what the user typed in the field.

Removed all the added code instead of just fixing this bug, because
its not a good way to solve masked input. If the user would edit an
input twice the result of templating/masking would be added to the
model value, not just control value. When we need masked input it
should be created properly.