fix(Uniform): Fix nullpointer and _initValue error in multivalue
Got nullpointer because getDisplayValue sometimes returned
null/undefined, fixed that by returning empty string if it was going to
return null/undefined.

The problem with the way _initValue was implemented was that it was
never removed from the object after creation. Fixed this by only
honoring _initValue when the field _initValue was intended for is empty.