feature(Bureau): Add a detail split view to the bureau page
Because we want to make specific requests for each company when somebody
clicks on the line of the company. But not when the click on the text
for the line, then we go to another view, this part might require some
re-thinking...

Please note that I added the splitting css/sass code as a standalone
common thing, to make it easier to re-use for other parts of the code.

Added a new http service so that we don't mix up the non auto company
key service with the one that does auto inject the company key. This
also means that we can have the error handling directly in the http
service since there are no other users of this service.

ref: #UC-153