fix(uniimage): use correct highlighting for OCR
the implemented solution is a hack that substracts the width/height from the left/top - probably because azure adds quite a lot of padding. However this causes problems when using PDF interpretation when OCR interpreting. Use the original top/left instead, this makes more sense normally (and we should also have a look at reducing the padding from cognitive services, but that can be handled later)