feat(unitable): support dropdowns in filters to support filtering for statuscodes
because using the statuscodes directly are not so user friendly, so the user should be able to select the statuscode by name from a dropdown. This filtering could also be used for other filtering, any field using predefined values can use it by setting filterOptions. In the ticker and query views, this is done automatically for StatusCode (valid statuses for the entity is also preloaded using the statusService)

closes UK-459