fix(loginModal): trying to fix bug in loginModal (see commit body)
The uni-form model in loginModal sometimes doesn't recognize input
from the password field. This seems to happen when the modal is
opened while computer is in sleep mode or when the modal
has been open for a long time (typically over night).
Can't reliably reproduce it when developing, so I'm just temporarily
removing this and seeing if it helps..