chore: revert previous change of moving rowSelect event to click
This was done because rowSelect event was designed to be used for
when the user clicks a row in a selection list, but has been used
for other stuff in the app. When I changed it some of these views
broke. Its a bad idea having rowSelect emit in focus because it
triggers on mousedown, not up, which can lead to some weird behavior
when navigation from selection lists, but whatever..