fix(Notifications): Error when click notification lost company access
Because we can't find the company when trying to open a notification we
get a crash and a stacktrace. The temporary fix for this is to show a
warning when the user tries to click a notification from a company they
no longer have access to. The long term fix is to filter out
notifications that the user doesn't have access to in the backend.

ref: UC-720
code-review: #3977