refactor(uniform): removed inline templates
 refactored uniform extracting templates. to improve component readability. 
Fixed lint errors as much as possible.

Also have tried to extract state of form and fields to centralise state but finally It has been seen that state is defined mainly by inputs and outputs and other fields used to represent the state are one, two or three, so it is not difficult to know what is the state of each component. Common code is mostly under BaseControl, which can be considered a kind of "state" class.