feature(License): Add license acceptance pop up
Because we want our users to agree to our license before using the
product.

Added check to not auto add button in modalService so that we can have more (any) control over what happens when the user clicks that. If the modal service adds the button the modal will close no matter what if the user clicks it. Now if we add it manually we can show a warning if the user clicks it.

ref: UC-721, UC-728
code-review: #4040