fix(License): Ignoring HasAgreededToLicense if AgreementId is 0
Because a user can't agree to a license if there are no licenses. And
AgreementId === 0 is an indicator of that.

This is something that should really be done in the backend. But since
we are short on time and we are going to do an emergency deploy tomorrow
we'll put it into the frontend. It might be implemented in the backend
before deploy but since we don't know for sure we will have to do it
here also.