fix(Core): Show error from BrowserStorageService before reload
We reload the app if we can't read something from localStorage because
we are unsure what state we leave the app in if we just continue. But
that also means that the errors are harder to see. You'll just see it as
a random reload. So therefore I'm changing it to return null if a
particular value from localStorage wasn't able to be parsed as JSON.

This also fixes the bug we have been having that after a login the login 
view is shown again, but with the navbar showing on top.

code-review: #4147